% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARQiv.R
\docType{package}
\name{ARQiv}
\alias{ARQiv}
\alias{ARQiv-package}
\title{ARQiv R* package: Preparing for and Completing Whole-Organism Screening at High-througput rates}
\description{
The ARQiv R* package includes functions that fall into two categories - those applied to
'Pre-screening Assay Optimization' and 'Compound Analysis'. The functions allow the user to
calculate background signal, determine sample size, run quality control tests, perform virtual
experiments to simulate compound efficacy - and finally, to perform compound analysis
during iterative drug screen cycles.
}

