% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{MULTI.REAL.TIME}
\alias{MULTI.REAL.TIME}
\title{Compound Analysis-Multiple}
\usage{
MULTI.REAL.TIME(Pos.file = NULL, Neg.file = NULL, multi.Drug.file = NULL,
  well.label.file = NULL, Roy.file = NULL, bg.ratio = NULL,
  dic.output = NULL)
}
\arguments{
\item{Pos.file}{A .xml file derived from your positive control condition, i.e., that with the highest value. This 'bookends' sets of 10 drug plates.}

\item{Neg.file}{A .xml file derived from your negative control condition, i.e., that with the lowest value. This 'bookends' sets of 10 drug plates.}

\item{multi.Drug.file}{Multiple .xml files of drug treatment plates for for comparison to bordering controls.}

\item{well.label.file}{A .csv denoting the identity of each well of a 96-well plate, by condition or concentration.}

\item{Roy.file}{A .xml file of raw signal values from a 96-well plate of non-transgenic Roy larvae.}

\item{bg.ratio}{Numerical minimal 'signal' cutoff value (arbitrary units) applied to isolate fluorescent/luminescent signal above this value.}

\item{dic.output}{Output directory}
}
\value{
multiple .png files which each include: Boxplot, SSMD, and Heatmap detailing a specific drug compound plate.
}
\description{
This function creates multiple .png files that each contain one Boxplot, SSMD, and Heatmap detailing the results of a specific drug compound plate.
}

