% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{ParseXML}
\alias{ParseXML}
\title{Signal Extraction}
\usage{
ParseXML(xmlfile = NULL)
}
\arguments{
\item{xmlfile}{A .xml file provided by the plate reader containing raw data from regional scanns of each well.}
}
\value{
Numerical 'signal' values (arbitrary units) for all individual wells which are applied to downstream analyses.
}
\description{
This function reads .xml input files, extracts and isolates the signal from individual wells of a 96-well plate. This function is applied to both control and compound plates.
}

