% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{REAL.TIME}
\alias{REAL.TIME}
\title{Compound Analysis}
\usage{
REAL.TIME(Pos.file = NULL, Neg.file = NULL, Drug.file = NULL,
  well.label.file = NULL, Roy.file = NULL, bg.ratio = NULL,
  dic.output = NULL)
}
\arguments{
\item{Pos.file}{A .xml file derived from your positive control condition, i.e., that with the highest value.}

\item{Neg.file}{A .xml file derived from your negative control condition, i.e., that with the lowest value.}

\item{Drug.file}{A .xml file derived from single drug treatment condition for comparison to controls.}

\item{well.label.file}{A .csv denoting the identity of each well of a 96-well plate, by condition or concentration.}

\item{Roy.file}{A .xml file of raw signal values from a 96-well plate of non-transgenic Roy larvae.}

\item{bg.ratio}{Numerical minimal 'signal' cutoff value (arbitrary units) applied to isolate fluorescent/luminescent signal above this value.}

\item{dic.output}{Output directory}
}
\value{
A .png file including: Boxplot, SSMD, and Heatmap detailing the drug compound plate.
}
\description{
This function creates a .png file that contains one Boxplot, SSMD, and Heatmap detailing the results of a drug compound plate.
}

