% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{SAMPLE}
\alias{SAMPLE}
\title{Sample Size Determination}
\usage{
SAMPLE(Pos.file = NULL, Neg.file = NULL, Roy.file = NULL,
  Excel.file = NULL, effect.size = NULL, bg.ratio = NULL,
  dic.output = NULL)
}
\arguments{
\item{Pos.file}{A .xml file derived from your positive control
condition, i.e., that with the highest value.}

\item{Neg.file}{A .xml file derived from your negative
control condition, i.e., that with the lowest value.}

\item{Roy.file}{A .xml file of raw signal values from a
96-well plate of non-transgenic Roy larvae.}

\item{Excel.file}{A .xlsx file derived manually containing
single columns of 'signal' from both positive and negative controls.}

\item{effect.size}{A simulated effect size for virtual
experiments (e.g., 0.75 is a compound that simulates 75 percents efficacy when compared to the positive control condition).}

\item{bg.ratio}{Numerical minimal 'signal' cutoff value
(arbitrary units) applied to isolate fluorescent/luminescent signal above this value.
Default value is 3946.}

\item{dic.output}{Output directory.Default is current working directory.}
}
\value{
SampleSize.csv A .csv file of minimum sample sizes required
to observe simulated effect sizes based on varied TypeI/TypeII errors.
}
\description{
This function calculates the minimum sample size required for
your assay with varying paremeters, such as simulated effect size
and TypeI/TypeII error rates. This can be completed using .XML or .CSV file formats.
}

