% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{SIGNAL}
\alias{SIGNAL}
\title{Background Signal Calculation}
\usage{
SIGNAL(Roy.file = NULL)
}
\arguments{
\item{Roy.file}{A .xml file of raw signal values from a 96-well plate of non-transgenic Roy larvae.}
}
\value{
bg.ratio Numerical minimal 'signal' cutoff value (arbitrary units) applied to isolate fluorescent/luminescent signal above this value.
}
\description{
This function calculates background from non-transgenic larvae. They identical to their transgenic counterparts with regard to pigmentation (e.g., Roy Orbison pigmentation mutants).
}

