% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{SSMD.QC}
\alias{SSMD.QC}
\title{SSMD_QC Calculation}
\usage{
SSMD.QC(Pos.file = NULL, Roy.file = NULL, bg.ratio = NULL,
  Neg.file = NULL, Excel.file = NULL, dic.output = NULL)
}
\arguments{
\item{Pos.file}{A .xml file derived from your positive control condition, i.e., that with the highest value.}

\item{Roy.file}{A .xml file of raw signal values from a 96-well plate of non-transgenic Roy larvae.}

\item{bg.ratio}{Numerical minimal 'signal' cutoff value (arbitrary units) applied to isolate fluorescent/luminescent signal above this value.}

\item{Neg.file}{A .xml file derived from your negative control condition, i.e., that with the lowest value.}

\item{Excel.file}{A .xlsx file derived manually containing single columns of 'signal' from both positive and negative controls.}

\item{dic.output}{Output directory.Default is current working directory.}

\item{effect.size}{A simulated effect size for virtual experiments (e.g., 0.75 is a compound that simulates 75 percents efficacy when compared to the positive control condition).}

\item{sampling.num}{The number of invidiual wells you will sample from a 96-well plate.}

\item{times}{The number of iterative cycles of the chosen sampling number.}
}
\value{
A .csv file with SSMD_EST estimated via random sampling (i.e., a compound with 75% efficacy when compared to the positive control will have an SSMD score of X).
}
\description{
This function calculates SSMD_EST of your chosen
assay parameters, using both positive and negative control
 data. This can be completed using .XML or .CSV file formats.
}

